+++ jablib/src/main/java/org/jabref/logic/importer/fetcher/transformers/AbstractQueryTransformer.java
@@ -142,7 +142,6 @@ public abstract class AbstractQueryTransformer {
     protected String handleYearRange(String yearRange) {
         parseYearRange(yearRange);
         if (endYear == Integer.MAX_VALUE) {
-            // invalid year range
             return yearRange;
         }
         StringJoiner resultBuilder = new StringJoiner(getLogicalOrOperator());